refactor: update ucan codec to use new DID generation method
