+++ crypto/mpc/base.go
